% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amigos}
\alias{amigos}
\title{Breve descripción física y sentimental de un grupo de amigos}
\format{
Un tibble con 6 filas y 6 variables agrupadas en 3 columnas:

\describe{
\item{nombre}{nombre de los amigos}
\item{fisico}{edad, altura y masa en forma de lista}
\item{sentimental}{estado civil e información de hijos(TRUE ó FALSE)}
}
}
\usage{
amigos
}
\description{
La base de datos contiene información de 6 amigos incluyendo su físico
e información sentimental.
}
\keyword{datasets}
