% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carlos_fam_g}
\alias{carlos_fam_g}
\title{Lista con datos precisos de la familia de Carlos}
\format{
Una lista que contiene 3 listas (grupos) cada una con las tres mismas variables
con inofrmación de la familia de carlos:

\describe{
\item{Padres}{lista con variables: relacion, nombre y edad}
\item{Hijos}{lista con variables: relacion, nombre y edad}
\item{Hermanos}{lista con variables: relacion, nombre y edad}
}
}
\usage{
carlos_fam_g
}
\description{
La lista contiene información correcta de la familia de un
personaje ficticio llamado carlos.
}
\keyword{datasets}
