% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos_2}
\alias{datos_2}
\alias{datos_3}
\alias{datos_4_a}
\alias{datos_4_b}
\title{Diseños tabulares traducido}
\format{
\describe{
\item{País}{Paises con reportes de tuberculosis}
\item{Año}{año de reporteCur}
\item{Casos}{número de casos del pais para detrminado año}
\item{Población}{población del pais para detrminado año}
\item{Tasa}{casos dividido entre población}
}

Los datos son un subconjunto de los datos contenidos en el Informe Mundial
sobre Tuberculosis de la Organización Mundial de la Salud.
}
\source{
\url{http://www.who.int/tb/country/data/download/en/}

Datos originales en R obtenidos en DSR::table1
}
\usage{
datos_2
}
\description{
Conjuntos de datos que demuestran múltiples formas de diseñar
los mismos datos tabulares.
}
\details{
datos_2, datos_3, datos_4_a, datos_4_b muestran el número de casos
de Tuberculosis (TB) documentados por la Organización Mundial de la Salud
en Afganistán, Brasil y China entre 1999 y 2000. Los datos contienen valores
asociados con cuatro variables (país, año,casos y población), pero cada tabla
organiza los valores en un diseño diferente.
}
\keyword{datasets}
