% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Datos_a}
\alias{Datos_a}
\alias{Datos_b}
\title{Datos destinados para usar con **_join()}
\format{
Dos tibble con una variable cada uno:

\describe{
\item{ID}{identificador}
\item{x1}{variable 1}
\item{x1}{variable 2}
}
}
\usage{
Datos_a
}
\description{
Conjuntos de datos para unir utilizando funciones **_join() del
tidyverse.
}
\keyword{datasets}
