% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departamentos}
\alias{departamentos}
\title{Datos de polígono para departamentos de Colombia}
\format{
Conjunto de información de homicidios en Colombia año 2013

\describe{
\item{long}{valores de longitud para polígono}
\item{lat}{valores de latitud para polígono}
\item{order}{orden de las lineas del poligono por grupo}
\item{hole}{TRUE:hueco; FALSE: no hueco}
\item{piece}{código de pieza en el dataset}
\item{group}{código indicativo de departamento}
\item{id}{id de departamento}
\item{depto}{nombre del departamento}
\item{lab_long}{valor de longitud para etiqueta de departamento}
\item{lab_lat}{valor de latitud para etiqueta de departamento}
\item{poblacion}{población de cada departamento en 2013}
\item{homicidios}{número de homicidios en 2013 para cada departamento}
\item{tasa}{número de homicidios por cada 100.000 habitantes}
}
}
\source{
\url{https://github.com/nebulae-co/colmaps}
\url{https://github.com/nebulae-co/homicidios}

Datos originales en R obtenidos en colmaps::departamentos y homicidios::homicidios
}
\usage{
departamentos
}
\description{
Los datos fueron obtenidos de dos paquetes \code{colmaps} y \code{homicidios} de
Daniel Moreno, contiene datos de homicidios en colombia del año
2013 para cada departamento incluyendo su población para la fecha.
}
\keyword{datasets}
