% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descuento_perros.R
\name{descuento_perros}
\alias{descuento_perros}
\title{Descuento en comida para perro}
\usage{
descuento_perros(bolsas, descuentos, to_vector = FALSE)
}
\arguments{
\item{bolsas}{número de bolsas compradas}

\item{descuentos}{código de descuento 1: hay descuento 0: no hay descuento}

\item{to_vector}{valor lógico que determina si la salida es vector o impresiones en pantalla}
}
\value{
impresiones en pantalla con \code{to_vector=FALSE}, vector con \code{to_vector=TRUE}
}
\description{
Esta función calcula el total a pagar en la compra de comida de perro,
si el descuento aplica es de un 50\% sobre el total de la compra y
se aplica un 10\% extra si el número de bolsas compradas es por lo menos 10.
El precio por bolsa de comida es de 6800 pesos.
}
\examples{
descuento_perros(c(10,2,15,18,16),c(1,0,0,1,0))

descuento_perros(c(10,2,15,18,16),c(1,0,0,1,0),to_vector=TRUE)

}
