% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england}
\alias{england}
\title{Resultados de la liga inglesa 1888-2017}
\format{
Un marco de datos con 196076 filas y 12 variables:
\describe{
\item{Date}{Fecha de coincidencia}
\item{Season}{Temporada de partido - se refiere al año de inicio}
\item{home}{Equipo local}
\item{visitor}{Equipo visitante}
\item{FT}{Resultado a tiempo completo}
\item{hgoal}{Goles marcados por el equipo local}
\item{vgoal}{Goles marcados por el equipo visitante}
\item{division}{División: 1,2,3,4 o 3a (Viejo 3-Norte) o 3b (Viejo 3-Sur)}
\item{tier}{Nivel de la pirámide de fútbol: 1,2,3,4}
\item{totgoal}{Total de goles en el juego}
\item{goaldif}{Diferencia de goles en los objetivos de inicio del juego - objetivos de visitante}
\item{result}{Resultado: H-Home Win, A-Away Win, D-Draw}
}
}
\source{
\url{https://github.com/jalapic/engsoccerdata}

Datos originales en R obtenidos en engsoccerdata::england
}
\usage{
england
}
\description{
Todos los resultados de los partidos de fútbol inglés en los 4 niveles superiores
desde la temporada 1888/89 hasta la temporada 2016/17
}
\keyword{datasets}
