% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{letras}
\alias{letras}
\alias{logicos}
\alias{numeros}
\title{Vectores de 3 clases para ejercicios}
\format{
Tres vectores separados:

\describe{
\item{letras}{contiene letras del abecedario de la a a la o}
\item{logicos}{contiene valores logicos}
\item{numeros}{contiene una secuencia de nuemors del 1 al 100}
}
}
\usage{
letras
}
\description{
Vectores cortos de la misma longitud (15) de 3 distintas clases
para crear data.frame y otros objetos.
}
\keyword{datasets}
