% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars_1}
\alias{mtcars_1}
\alias{mtcars_2}
\alias{mtcars_3}
\title{Subconjuntos de pruebas de carretera de automóvil de motor trend}
\format{
\describe{
\item{cyl}{número de cilindros}
\item{hp}{potencia bruta}
\item{wt}{peso}
\item{am}{transmisión}
\item{carb}{número de carburadores}
}

Los datos son un subconjunto de mtcars
}
\source{
Henderson and Velleman (1981), Building
multiple regression models interactively. Biometrics, 37, 391–411.

Datos originales en R obtenidos en datasets::mtcars
}
\usage{
mtcars_1
}
\description{
Los datos se extrajeron de la revista Motor Trend US de 1974 y
comprenden el consumo de combustible y 10 aspectos del diseño y
rendimiento de automóviles para 32 automóviles (modelos 1973-74).
}
\keyword{datasets}
