% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_lista}
\alias{my_lista}
\title{lista con distintos tipos de objetos}
\format{
\describe{
\item{vector}{vector con números del 1 al 10}
\item{matriz}{matriz 5X5 de los números del 1 al 25}
\item{dataframe}{Primeras 10 observaciones de base de datos Starwars}
\item{tible}{mtcars_3 convertido a objeto tibble}
}

La lista contiene dos objetos creados recursivamente y 2  objetos modificados
especialmente para este paquete como lo son Starwars y mtcars_3.
}
\usage{
my_lista
}
\description{
lista con 4 objetos, un vector, una matriz, un dataframe y un tibble
}
\details{
Datos en R obtenidos en CursoR::Starwars , CursoR::Starwars. Para una
descripción detallada consultelos individualmente
}
\keyword{datasets}
