% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokemon}
\alias{pokemon}
\alias{pokemon_2}
\title{Datos sobre Pokémon}
\format{
Conjunto de información de 800 pokémon

\describe{
\item{id}{ID de cada pokémon (numérica)}
\item{nombre}{Nombre en español (caracter)}
\item{tipo}{Tipo de pokémon (caracter)}
\item{tipo_2}{Algunos pokémon son de dos tipos. Para el resto, esta variable aparece como NA (caracter)}
\item{ataque}{Daño en ataques normales (numérica)}
\item{defensa}{Resistencia al daño ante ataques normales (numérica)}
\item{velocidad}{Determina qué pokémon ataca primero en cada ronda (numérica)}
\item{puntos_vida}{Cantidad de daño que puede resistir (numérica)}
}
}
\source{
\url{https://github.com/cienciadedatos/datos-de-miercoles/tree/master/datos/2019/2019-07-10}
}
\usage{
pokemon
}
\description{
Los datos fueron obtenidos desde Kaggle y la Pokémon Database.
El dataset fue propuesto, traducido y editado por Gonzalo Fichero para
datos de miércoles semana 14 y se adapto al paquete.El dataset contiene
datos sobre las características de 800 pokémon.
}
\keyword{datasets}
