% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pts_temporada}
\alias{Pts_temporada}
\title{Datos sobre Pokémon}
\format{
Un marco de datos con 2443 filas y 3 variables:

\describe{
\item{Temporada}{Año en el que se llevo a cabo la competición 1888-2017}
\item{Equipo}{Nombre del los equipos pertenecientes a la primera división}
\item{Puntos}{Puntos logrados por el respectivo equipo en esa temporada de futbol}
}

Datos obtenidos a partir de engsoccerdata::england
}
\usage{
Pts_temporada
}
\description{
Los datos fueron obtenidos desde Kaggle y la Pokémon Database.
El dataset fue propuesto, traducido y editado por Gonzalo Fichero para
datos de miércoles semana 14 y se adapto al paquete.El dataset contiene
datos sobre las características de 800 pokémon.
}
\keyword{datasets}
