% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Starwars}
\alias{Starwars}
\title{Personajes de Starwars}
\format{
Un dataframe modificado con 87 filas y 11 variables:

\describe{
\item{name}{nombre del personaje}
\item{height}{altura (cm)}
\item{mass}{masa (kg)}
\item{hair_color}{color de cabello}
\item{skin_color}{color de piel}
\item{eye_color}{color de ojos}
\item{birth_year}{año de nacimiento (Antes de la batalla de Yavin)}
\item{sex}{el sexo biológico del personaje, es decir, masculino, femenino, hermafrodita o ninguno (como en el caso de los droides)}
\item{gender}{el rol de género o la identidad de género del personaje según lo determinado por su personalidad o la forma en que fueron programados (como en el caso de los droides)}
\item{homeworld}{nombre del mundo natal}
\item{species}{nombre de la especie}
}

Los datos son un subconjunto de los datos starwars provenientes del paquete dplyr.
}
\source{
\url{http://swapi.dev/}

Datos originales en R obtenidos en dplyr::starwars
}
\usage{
Starwars
}
\description{
Estos datos provienen de SWAPI, la API de Star Wars, \url{http://swapi.dev/}
}
\keyword{datasets}
