% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tuberculosis}
\alias{Tuberculosis}
\title{Datos de TB de la Organización Mundial de la Salud}
\format{
Un conjunto de datos con 7240 observaciones y 60 variables.

\describe{
\item{country}{Paises con reportes de tuberculosis}
\item{iso2}{Código de país de dos letras de la Organización Internacional de Normalización (ISO)}
\item{iso3}{Código de país de tres letras de la Organización Internacional de Normalización (ISO)}
\item{new_sp_m014 - new_rel_f65}{Recuentos de nuevos casos de TB registrados por grupo. Los nombres de columna codifican tres variables que describen el grupo}
}
}
\source{
\url{http://www.who.int/tb/country/data/download/en/}

Datos originales en R obtenidos en DSR::who
}
\usage{
Tuberculosis
}
\description{
Un subconjunto de datos del Informe Mundial de Tuberculosis de
la Organización Mundial de la Salud.
}
\details{
Los datos utilizan los códigos originales proporcionados por la Organización
Mundial de la Salud. Los nombres de columna para las columnas cinco a 60 se
crean combinando new_ a un código para el método de diagnóstico (rel = recaída,
sn = frotis pulmonar negativo, sp = frotis pulmonar positivo, ep = extrapulmonar)
a un código de género (f = hembra , m = masculino) a un código para el grupo
de edad (014 = 0-14 años de edad, 1524 = 15-24 años de edad, 2534 = 25 a 34
años de edad, 3544 = 35 a 44 años de edad, 4554 = 45 a 54 años, 5564 = 55 a
64 años, 65 = 65 años o más).
}
\keyword{datasets}
