% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USArrests}
\alias{USArrests}
\title{Datos de polígono para estados de USA}
\format{
Conjunto de información de homicidios en Colombia año 2013

\describe{
\item{region}{nombre del estado}
\item{murder}{Detenciones por asesinato (por cada 100,000)}
\item{Assault}{Detenciones de asalto (por 100,000)}
\item{UrbanPop}{porcentaje de población urbana}
\item{Rape}{Detenciones por violación (por 100,00)}
\item{long}{valores de longitud para polígono}
\item{lat}{valores de latitud para polígono}
\item{group}{código indicativo de departamento}
\item{order}{orden de las lineas del poligono por grupo}
\item{subregion}{subregion}
}
}
\source{
Datos originales en R obtenidos en datasets::USArrests y ggplot2::map_data("state")
}
\usage{
USArrests
}
\description{
Este conjunto de datos contiene estadísticas sobre arrestos por cada 100,000
residentes por asalto, asesinato y violación en cada uno de los 50 estados
de EE. UU. En 1973. También se da el porcentaje de la población que
vive en áreas urbanas y se proporciona información de polígono para
gráfico.
}
\keyword{datasets}
