% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CursoR.R
\docType{package}
\name{CursoR}
\alias{CursoR}
\title{Un Curso Amigable Para R}
\description{
Contiene un Curso amigable para R enfocado especialmente en el tidyverse
}
\author{
Leonardo Lopez \email{jollopezma@unal.edu.co} & Duvan Nieves \email{dnieves@unal.edu.co}
}
